requirements.txt
Pillow>=9.0.0
streamlit>=1.20.0 